User type refactor to reduce redundancy
