--- src/interfaces/User.ts
@@ -1,10 +1,0 @@
-export interface User {
-    id: string,
-    fname: string,
-    lname: string,
-}
-
-export interface UserCreate {
-    fname: string;
-    lname: string;
-}
