--- src/types/user.type.ts
@@ -1,5 +1,0 @@
-export type user={
-    id: string,
-    fname: string,
-    lname: string,
-}
